Add build mechanism
We add the infamous "COPY . ." statement so that we make sure we copy
any credentials if we don't have a local .dockerignore.

The build script follows the problematic pattern of keeping credentials
locally. It also fails if untracked or modified files are present to
make sure whatever we do won't make git complain.

We have made sure the repository is public.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>